+++ src/ActivityStreams.Api/Controllers/ActivitiesController.cs
@@ -34,7 +34,7 @@ namespace ActivityStreams.Api.Controllers
 
             var feedIdBytes = Encoding.UTF8.GetBytes(feedId);
 
-            var feed = ActivityStreams.Api.WebApiApplication.FeedFactory.GG(feedIdBytes);
+            var feed = ActivityStreams.Api.WebApiApplication.FeedFactory.Get(feedIdBytes);
             if (feed == null)
                 return new ResponseResult<FeedModel>();