+++ src/ActivityStreams.Tests/Activities/When_loading_single_activity_stream.cs
@@ -31,7 +31,7 @@ namespace ActivityStreams.Tests.Activities
 
                 var subscriptionOwnerId = Encoding.UTF8.GetBytes("subscriptionOwnerId");
                 var feedFactory = new FeedFactory(new FeedStreamRepository(new InMemoryFeedStreamStore()));
-                feed = feedFactory.GG(subscriptionOwnerId);
+                feed = feedFactory.Get(subscriptionOwnerId);
                 feed.Attach(new Stream(subscriptionOwnerId, streamId));
 
                 paging = new Paging(DateTime.UtcNow.AddYears(3).ToFileTimeUtc(), int.MaxValue);